Add Speech to Text service
This commit adds Speech to Text capabilities to the chat application. In particular, it streams microphone audio to the Speech to Text service when the microphone button is touched. Interim transcription results from the service are used to set the text field as they are received. Streaming ends when the microphone button is released by the user.